+++ CAPS/Core/Service/NavigationService.cs
@@ -4,22 +4,22 @@ namespace CAPS.Core.Service;
 
 public interface INavigationService
 {
-	ViewModel CurrentView { get; }
-	void NavigateTo<T>() where T : ViewModel;
+	ViewModelBase CurrentView { get; }
+	void NavigateTo<T>() where T : ViewModelBase;
 }
 
 public partial class NavigationService : ObservableObject, INavigationService
 {
 	[ObservableProperty]
-	private ViewModel currentView;
-	private Func<Type, ViewModel> _viewModelFactory;
+	private ViewModelBase currentView;
+	private Func<Type, ViewModelBase> _viewModelFactory;
 
-	public NavigationService(Func<Type, ViewModel> viewModelFactory)
+	public NavigationService(Func<Type, ViewModelBase> viewModelFactory)
     {
         _viewModelFactory = viewModelFactory;
     }
 
-    public void NavigateTo<TViewModel>() where TViewModel : ViewModel
+    public void NavigateTo<TViewModel>() where TViewModel : ViewModelBase
 	{
 		CurrentView = _viewModelFactory.Invoke(typeof(TViewModel));
 	}
