+++ CAPS/View/MainWindow.xaml.cs
