+++ CAPS/View/MissionPlanning/WaypointList.xaml
