+++ CAPS/View/MissionPlanning/WaypointList.xaml.cs
