+++ CAPS/ViewModel/MainWindowViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using CAPS.Core;
 using CAPS.Services.Mission;
-using CAPS.Views.Components;
+using CAPS.ViewModel.Framework;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using System.Collections.ObjectModel;
@@ -8,7 +8,7 @@ using System.Reflection;
 
 namespace CAPS.Views;
 
-public partial class MainWindowViewModel : ViewModel
+public partial class MainWindowViewModel : ViewModelBase
 {
 	public ObservableCollection<TreeViewItemViewModel> TreeViewItems { get; }
